// src/types/github-graphql.ts
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string,
  String: string,
  Boolean: boolean,
  Int: number,
  Float: number,
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: any,
  /** An ISO-8601 encoded UTC date string. */
  DateTime: any,
  /** A string containing HTML code. */
  HTML: any,
  /** A Git object ID. */
  GitObjectID: any,
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: any,
  /** Git SSH string */
  GitSSHRemote: any,
  /** An ISO-8601 encoded date string. */
  Date: any,
  /** An ISO-8601 encoded UTC date string with millisecond precison. */
  PreciseDateTime: any,
  /** A valid x509 certificate string */
  X509Certificate: any,
  /** A fully qualified reference name (e.g. `refs/heads/master`). */
  GitRefname: any,
};
// ...
export enum OrderDirection {
  /** Specifies an ascending order for a given `orderBy` argument. */
  Asc = 'ASC',
  /** Specifies a descending order for a given `orderBy` argument. */
  Desc = 'DESC'
}
// ...
export type PageInfo = {
   __typename?: 'PageInfo',
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['String']>,
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean'],
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean'],
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['String']>,
};
// ...
export type StarOrder = {
  /** The direction in which to order nodes. */
  direction: OrderDirection,
  /** The field in which to order nodes by. */
  field: StarOrderField,
};

/** Properties by which star connections can be ordered. */
export enum StarOrderField {
  /** Allows ordering a list of stars by when they were created. */
  StarredAt = 'STARRED_AT'
}
// ...
export type UserStarredRepositoriesArgs = {
  after?: Maybe<Scalars['String']>,
  before?: Maybe<Scalars['String']>,
  first?: Maybe<Scalars['Int']>,
  last?: Maybe<Scalars['Int']>,
  orderBy?: Maybe<StarOrder>,
  ownedByViewer?: Maybe<Scalars['Boolean']>
};